fix: shared RemOptions
